cloud_function_producer/requirements.txt
gtfs-realtime-bindings==1.0.0
kafka-python==2.0.2
requests==2.28.2